+++ src/Trace.hs
@@ -6,29 +6,43 @@
 {-# LANGUAGE TypeOperators #-}
 {-# LANGUAGE AllowAmbiguousTypes #-}
 {-# LANGUAGE FlexibleInstances #-}
+{-# LANGUAGE PatternSynonyms #-}
 {-# LANGUAGE UndecidableInstances #-}
 {-# OPTIONS_GHC -Wno-incomplete-patterns #-}
 
-module Trace where
+
+{- | For recording samples and log-probabilities during model execution 
+-}
+
+module Trace (
+  -- * Sample trace
+    STrace
+  , FromSTrace(..)
+  , filterSTrace
+  , traceSamples
+  -- * Log-probability trace
+  , LPTrace
+  , traceLPs) where
 
 import Data.Map (Map)
-import Data.Maybe
-import Data.Proxy
-import Effects.Dist
-import PrimDist
-import Env
-import GHC.TypeLits
+import Data.Maybe ( fromJust )
+import Data.Proxy ( Proxy(..) )
+import Effects.Dist ( Tag, Addr, Observe, Sample(..), pattern ObsPrj, pattern SampPrj )
+import Env ( enil, varToStr, UniqueVar, Var(..), Env(ECons), Assign((:=)) )
+import GHC.TypeLits ( KnownSymbol )
 import OpenSum (OpenSum)
 import qualified Data.Map as Map
 import qualified OpenSum
-import Prog
-import Effects.State
+import Prog ( Member, Prog(..), weaken, install )
+import PrimDist ( ErasedPrimDist(..), PrimVal, PrimDist, logProb, pattern PrimDistPrf )
+import Effects.State ( State, modify, handleState )
 
--- ||| Sample trace, mapping addresses of sample/observe operations to their primitive distributions and sampled values
+-- | The type of sample traces, mapping addresses of sample/observe operations to their primitive distributions and sampled values
 type STrace = Map Addr (ErasedPrimDist, OpenSum PrimVal)
 
--- ||| For converting sample traces, as used by simulation and inference, to output model environments
+-- | For converting sample traces, as used by simulation and inference, to output model environments
 class FromSTrace a where
+  -- | Convert a sample trace to a model environment
   fromSTrace :: STrace -> Env a
 
 instance FromSTrace '[] where
@@ -36,20 +50,34 @@ instance FromSTrace '[] where
 
 instance (UniqueVar x env ~ 'True, KnownSymbol x, Eq a, OpenSum.Member a PrimVal, FromSTrace env) => FromSTrace ((x := a) : env) where
   fromSTrace sMap = ECons (extractSTrace (Var @x, Proxy @a) sMap) (fromSTrace sMap)
+    where extractSTrace ::  forall a x. (Eq a, OpenSum.Member a PrimVal) => (Var x, Proxy a) -> STrace -> [a]
+          extractSTrace (x, typ)  =
+              map (fromJust . OpenSum.prj @a . snd . snd)
+            . Map.toList
+            . Map.filterWithKey (\(tag, idx) _ -> tag == varToStr x)
 
-extractSTrace ::  forall a x. (Eq a, OpenSum.Member a PrimVal) => (Var x, Proxy a) -> STrace -> [a]
-extractSTrace (x, typ)  =
-    map (fromJust . OpenSum.prj @a . snd . snd)
-  . Map.toList
-  . Map.filterWithKey (\(tag, idx) _ -> tag == varToStr x)
-
-filterSTrace :: [Tag] -> STrace -> STrace
+-- | Retrieve the sampled values for the specified observable variable names
+filterSTrace :: 
+  -- | Observable variable names
+    [Tag]  
+  -- | Sample trace
+  -> STrace 
+  -- | Filtered sample trace
+  -> STrace
 filterSTrace tags = Map.filterWithKey (\(tag, idx) _ -> tag `elem` tags)
 
-updateSTrace :: Show x => (Member (State STrace) es, OpenSum.Member x PrimVal) => Addr -> PrimDist x -> x -> Prog es ()
+-- | Update a sample trace at an address
+updateSTrace :: (Show x, Member (State STrace) es, OpenSum.Member x PrimVal) => 
+  -- | Address of sample site
+     Addr 
+  -- | Primitive distribution at address
+  -> PrimDist x 
+  -- Sampled value
+  -> x 
+  -> Prog es ()
 updateSTrace α d x  = modify (Map.insert α (ErasedPrimDist d, OpenSum.inj x) :: STrace -> STrace)
 
--- ||| Trace sampled values for each Sample operation
+-- | Insert stateful operations for recording the sampled values at each @Sample@ operation
 traceSamples :: (Member Sample es) => Prog es a -> Prog es (a, STrace)
 traceSamples = handleState Map.empty . storeSamples
   where storeSamples :: (Member Sample es) => Prog es a -> Prog (State STrace ': es) a
@@ -59,14 +87,22 @@ traceSamples = handleState Map.empty . storeSamples
                                              k x
           )
 
--- ||| Log probability trace, mapping addresses of sample/observe operations to their log probabilities
+-- | The type of log-probability traces, mapping addresses of @Sample@ and @Observe@ operations to their log probabilities
 type LPTrace = Map Addr Double
 
-updateLPTrace :: (Member (State LPTrace) es) => Addr -> PrimDist x -> x -> Prog es ()
+-- | Compute and update a log-probability trace at an address
+updateLPTrace :: (Member (State LPTrace) es) => 
+  -- | Address of sample/observe site
+    Addr 
+  -- | Primitive distribution at address
+  -> PrimDist x 
+  -- | Sampled or observed value
+  -> x 
+  -> Prog es ()
 updateLPTrace α d x  = modify (Map.insert α (PrimDist.logProb d x) :: LPTrace -> LPTrace)
 
--- ||| Insert stateful operations for LPTrace when either Sample or Observe occur.
-traceLPs ::(Member Sample es, Member Observe es) => Prog es a -> Prog es (a, LPTrace)
+-- | Insert stateful operations for recording the log-probabilities at each @Sample@ or @Observe@ operation
+traceLPs :: (Member Sample es, Member Observe es) => Prog es a -> Prog es (a, LPTrace)
 traceLPs = handleState Map.empty . storeLPs
   where storeLPs :: (Member Sample es, Member Observe es) => Prog es a -> Prog (State LPTrace: es) a
         storeLPs (Val x) = pure x